% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependemcies.R
\name{aa_count}
\alias{aa_count}
\title{aa_count}
\usage{
aa_count(Amino_acid)
}
\arguments{
\item{Amino_acid}{A character string containing a single amino acid sequence.}

\item{Unique_aa}{Set of unique amino acids present}
}
\value{
A ggplot object showing the frequency of each amino acid in the input sequence.
}
\description{
aa_count
}
\examples{
aa_count("ACDEFGHIKLMNPQRSTVWY")
aa_count("AAAAAGGGCCCTTT")
}
