% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_rna.R, R/hello.R
\name{convert_to_rna}
\alias{convert_to_rna}
\title{Convert DNA sequence to RNA sequence}
\usage{
convert_to_rna(dna_sequence)

convert_to_rna(dna_sequence)
}
\arguments{
\item{dna_sequence}{The nucleotide composition of the DNA sequence intended for the function.}
}
\value{
RNA sequence of converted DNA sequence. Note the RNA sequence is NOT the sequence complementary to the DNA sequence.

The RNA sequence converted from the DNA sequence. Note: The generated RNA sequence is NOT the complementary RNA sequence.
}
\description{
Convert DNA sequence to RNA sequence

Generate RNA sequence from DNA sequence
}
\examples{
dna1 <- c("AAATTTCCTTGG")
rna1 <- convert_to_rna(dna_sequence = dna1)
print(rna1)


dna2 <- c("AATTCCTTCCTT")
rna2 <- convert_to_rna(dna_sequence = dna2)
print(rna2)


dna1 <- c("ACCTCTGCCACTTTACG")
          convert_to_rna(dna_sequence = dna1)

          dna2 <- c("AAACCCTTTCCCTTGGGCCTTT")
          convert_to_rna(dna_sequence = dna2)

}
