% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dna.R
\name{create_dna}
\alias{create_dna}
\title{Generate a Random DNA sequence}
\usage{
create_dna(length)
}
\arguments{
\item{length}{Specify the length of the dna}
}
\value{
A character string
}
\description{
Creates a single string that represents a random DNA sequence given a specific length
}
\examples{
create_dna(length=15)
}
