% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2_dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Generate RNA sequence from DNA sequence}
\usage{
dna_to_rna(dna_sequence)
}
\arguments{
\item{dna_sequence}{The nucleotide composition of the DNA sequence intended for the function.}
}
\value{
The RNA sequence converted from the DNA sequence. Note: The generated RNA sequence is NOT the complementary RNA sequence.
}
\description{
Generate RNA sequence from DNA sequence
}
\examples{
dna1 <- c("ACCTCTGCCACTTTACG")
          dna_to_rna(dna_sequence = dna1)

          dna2 <- c("AAACCCTTTCCCTTGGGCCTTT")
          dna_to_rna(dna_sequence = dna2)

}
