% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genetic_code}
\alias{genetic_code}
\title{Genetic Code}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://r4bds.github.io/lab08.html}
}
\usage{
genetic_code
}
\description{
All codons are linked to their corresponding aminoacid
}
\keyword{datasets}
