% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucMerge.R
\name{nucMerge}
\alias{nucMerge}
\title{Merging nucleotides}
\usage{
nucMerge(number)
}
\arguments{
\item{number}{How many nucleotides should be collapsed together}
}
\value{
Merged sequence from different nucleotides
}
\description{
Merging nucleotides
}
