% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAACount.R
\name{PlotAACount}
\alias{PlotAACount}
\title{Plot Amino Acid Counts}
\usage{
PlotAACount(AA)
}
\arguments{
\item{AA}{A string of Amino acids}
}
\value{
The function returns a barplot of Amino Acid Counts
}
\description{
Plot Amino Acid Counts
}
