% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f3_rna_to_codons.R
\name{r2c}
\alias{r2c}
\title{Title: Convert your RNA string into sets of codons}
\usage{
r2c(rna_strand, start = 1)
}
\arguments{
\item{rna_strand}{This is the RNA sequence}

\item{start}{=1 This is by default}
}
\value{
Sets of Codons
}
\description{
Title: Convert your RNA string into sets of codons
}
\examples{
r2c("ABCDEFGHIJKL")
}
