% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_nt_seq.R
\name{sample_nt_seq}
\alias{sample_nt_seq}
\title{Title: Sampeling nucleotide sequence}
\usage{
sample_nt_seq(seq_length = 99)
}
\arguments{
\item{seq_length}{desired length of seqence}
}
\value{
randomly generates nucleotide sequence of specified length
}
\description{
Title: Sampeling nucleotide sequence
}
\examples{
sample_nt_seq(6)  # Expected output: "GTAATT"


}
