% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqToCodon.R
\name{seqToCodons}
\alias{seqToCodons}
\title{The seqToCodons-function}
\usage{
seqToCodons(sequence1, start = 1)
}
\arguments{
\item{sequence1}{sequence1 is the nucleotide sequence provided by the user.}

\item{start}{start is the first nucleotide in the sequence.}
}
\value{
........doing it later
}
\description{
The seqToCodons-function
}
