% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_codons.R
\name{split_codons}
\alias{split_codons}
\title{Split a nucleotide sequence into codons}
\usage{
split_codons(amino, start = 1)
}
\arguments{
\item{amino}{character string representing the nucleotide sequence to split}

\item{start}{integer specifying the starting position for reading codons (default is 1)}
}
\value{
character vector containing codons (triplets of nucleotides)
}
\description{
Split a nucleotide sequence into codons
}
\examples{
# Example 1: Standard reading frame
split_codons("ATGGCCATTGTAATGGGCCGCTGAAAGGGTGCCCGATAG")

# Example 2: Starting from the second nucleotide
split_codons("ATGGCCATTGTAATGGGCCGCTGAAAGGGTGCCCGATAG", start = 2)

}
