% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_protein.R
\name{to_protein}
\alias{to_protein}
\title{to_protein}
\usage{
to_protein(codons)
}
\arguments{
\item{codons}{Three-letter sequences of RNA}
}
\value{
It returns the translated protein
}
\description{
to_protein
}
\examples{
to_protein(c("ACU", "UUU", "UGG", "UAA"))
}
