% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rna.R
\name{to_rna}
\alias{to_rna}
\title{Convert DNA sequence to RNA}
\usage{
to_rna(dna)
}
\arguments{
\item{dna}{A character string representing a DNA sequence (e.g., "ATGC").}
}
\value{
A character string representing the corresponding RNA sequence, where all "T" are replaced by "U".
}
\description{
This function takes a DNA sequence as input and converts it into the corresponding RNA sequence
by replacing all occurrences of thymine ("T") with uracil ("U").
}
\examples{
to_rna("ATGC")
to_rna("TTTT")
to_rna("GATTACA")
}
