% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_transl_RNA_2_AA.R
\name{transl_RNA_2_AA}
\alias{transl_RNA_2_AA}
\title{Title: Translation of RNA to AA}
\usage{
transl_RNA_2_AA(input_codons)
}
\arguments{
\item{input_codons, }{String vector with codons, separated by a comma (e.g. ATG, GCC etc.)}
}
\value{
Returns corresponding amino acid sequence as a collapsed continuous string
}
\description{
Title: Translation of RNA to AA
}
\examples{
input_vector <-  c("ATG", "GCC", "UGC", "AAG", "UGG", "CGT")
transl_RNA_2_AA(input_vector)
}
