% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranslateCodon.R
\name{TranslateCodon}
\alias{TranslateCodon}
\title{Translate Codon to Amino Acid}
\usage{
TranslateCodon(codons)
}
\arguments{
\item{codons}{A string of three characters specifying a codon}
}
\value{
The function returns the corresponding Amino Acid in one letter format
}
\description{
Translate Codon to Amino Acid
}
