% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranslateDNA.R
\name{TranslateDNA}
\alias{TranslateDNA}
\title{Translate DNA to RNA}
\usage{
TranslateDNA(DNAseq)
}
\arguments{
\item{DNAseq}{A DNA string consisting of nucleotides (A, T, C and G)}
}
\value{
The function returns the translated RNA string where all T's has been
substituted for U's.
}
\description{
Translate DNA to RNA
}
