% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_to_AA.R
\name{translation_RNA_2_AA}
\alias{translation_RNA_2_AA}
\title{translation_RNA_2_AA}
\usage{
translation_RNA_2_AA(codons_input)
}
\arguments{
\item{codons_input, }{a vector of characters split into three bases (e.g. ATG, GCC etc.)}
}
\value{
returns the amino acid sequence as a collapsed, continuous string,
for the given input sequence
}
\description{
translation_RNA_2_AA
}
\examples{
input_vector <- c(ATGGCCUGCAAGUGGCGT)
translation_RNA_2_AA(input_vector)
}
