% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f5_plot_aa_freqs.R
\name{plot_aa_freqs}
\alias{plot_aa_freqs}
\title{Title: Plotting amine acid frequency}
\usage{
plot_aa_freqs(aa_sequence)
}
\arguments{
\item{aa_sequence}{an amino acid seqence of any length.}
}
\value{
returns colored box plot with counts for each appearing amino acid.
}
\description{
Title: Plotting amine acid frequency
}
\examples{
plot_aa_freqs("AGSLYVDHK")

}
